+++ CHANGELOG.md
@@ -1,4 +1,10 @@
 # 1.x branch
+## 1.6 branch
+### 1.6.0
+* requires at least PHP 8.1, CakePHP 5 and PHPUnit 10;
+* fixed typehints;
+* added tests for PHP 8.3.
+
 ## 1.5 branch
 ### 1.5.13
 * several small code and test improvements;